Trail A is now organized in a smarter, more localized way.
Rather than having the trail traverse radomly across the page (making it harder
for the participant to find the 'next' circle), the trail now appears mostly in
a 'natural' ordering.

This was accomplished by enforcing an ordering on circles. For each circle, the
next circle is defined as its closest neighbour which is not already part of the
trail.

** note: I've prematurely commited a Comparator for Trail B as well. This should
be tested once appropriate.